refactor(reservation-app): Room 유스케이스 테스트 통합
- 유스케이스 별로 분산된 테스트 코드를 하나의 Command UseCase Test로 통합
